@@ -53,37 +30,29 @@
   "bugs": {
     "url": "https://github.com/strapi/strapi/issues"
   },
+  "bundleDependencies": false,
   "dependencies": {
     "sendmail": "^1.2.0"
   },
+  "deprecated": false,
   "description": "This is the description of the plugin.",
   "devDependencies": {
-    "strapi-helper-plugin": "3.0.0-alpha.11.3"
-  },
-  "directories": {},
-  "dist": {
-    "integrity": "sha512-70uLre/AmE9pD840t45WQ08n7QcoO394IZLplQIdAz6+n5SWcsH/QHcUR4pk9oor7a6MNmXbjZ0R3txly6E3Cw==",
-    "shasum": "79895a54c340ced6710307275f801c51c69d390c",
-    "tarball": "https://registry.npmjs.org/strapi-plugin-email/-/strapi-plugin-email-3.0.0-alpha.11.3.tgz",
-    "fileCount": 30,
-    "unpackedSize": 407953
+    "strapi-helper-plugin": "3.0.0-alpha.12.3"
   },
   "engines": {
-    "node": ">= 7.0.0",
-    "npm": ">= 3.0.0"
+    "node": ">= 9.0.0",
+    "npm": ">= 5.0.0"
   },
   "homepage": "https://github.com/strapi/strapi#readme",
   "license": "MIT",
   "maintainers": [
     {
-      "name": "strapi",
-      "email": "support@strapi.io"
+      "name": "Strapi team",
+      "email": "hi@strapi.io",
+      "url": "http://strapi.io"
     }
   ],
   "name": "strapi-plugin-email",
-  "optionalDependencies": {},
-  "readme": "# Strapi plugin",
-  "readmeFilename": "README.md",
   "repository": {
     "type": "git",
     "url": "git://github.com/strapi/strapi.git"
