@@ -81,8 +81,8 @@ export class EditPage extends React.Component {
     this.props.setFileRelations(fileRelations);
   }
 
-  componentWillReceiveProps(nextProps) {
-    if (nextProps.editPage.submitSuccess !== this.props.editPage.submitSuccess) {
+  componentDidUpdate(prevProps) {
+    if (prevProps.editPage.submitSuccess !== this.props.editPage.submitSuccess) {
       if (!isEmpty(this.props.location.search) && includes(this.props.location.search, '?redirectUrl')) {
         const redirectUrl = this.props.location.search.split('?redirectUrl=')[1];
 
@@ -118,7 +118,6 @@ export class EditPage extends React.Component {
    */
   getAttributeValidations = (name) => get(this.props.editPage.formValidations, [findIndex(this.props.editPage.formValidations, ['name', name]), 'validations'], {})
 
-
   /**
    * Retrieve the model
    * @type {Object}
@@ -130,7 +129,6 @@ export class EditPage extends React.Component {
    * @type {String} name
    */
   getModelAttribute = (name) => get(this.getModelAttributes(), name);
-
 
   /**
    * Retrieve the model's attributes
@@ -151,7 +149,6 @@ export class EditPage extends React.Component {
   getSchema = () => this.getSource() !== 'content-manager' ?
     get(this.props.schema, ['plugins', this.getSource(), this.getModelName()])
     : get(this.props.schema, [this.getModelName()]);
-
 
   /**
    * Retrieve the model's source
@@ -212,11 +209,6 @@ export class EditPage extends React.Component {
     }
 
     this.props.setFormErrors(formErrors);
-  }
-
-  componentDidCatch(error, info) {
-    console.log('err', error);
-    console.log('info', info);
   }
 
   isCreating = () => this.props.match.params.id === 'create';
