@@ -6,22 +6,25 @@
 
 import React from 'react';
 import { defineMessages, FormattedMessage } from 'react-intl';
-
-import LocaleToggle from 'containers/LocaleToggle';
+import { PropTypes } from 'prop-types';
 
 import styles from './styles.scss';
 import messages from './messages.json';
 defineMessages(messages);
 
-class LeftMenuFooter extends React.Component { // eslint-disable-line react/prefer-stateless-function
-  render() {
-    return (
-      <div className={styles.leftMenuFooter}>
-        <FormattedMessage {...messages.poweredBy} /> <a href="http://strapi.io" target="_blank">Strapi</a>
-        <LocaleToggle />
+function LeftMenuFooter({ version }) { // eslint-disable-line react/prefer-stateless-function
+  return (
+    <div className={styles.leftMenuFooter}>
+      <div>
+        <FormattedMessage {...messages.poweredBy} />
+        <a href="https://strapi.io" target="_blank">v{version}</a>
       </div>
-    );
-  }
+    </div>
+  );
 }
+
+LeftMenuFooter.propTypes = {
+  version: PropTypes.string.isRequired,
+};
 
 export default LeftMenuFooter;
