@@ -7,28 +7,31 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { FormattedMessage } from 'react-intl';
-import { get, snakeCase, isEmpty, map, sortBy } from 'lodash';
+import { findIndex, get, snakeCase, isEmpty, map, sortBy } from 'lodash';
 
 import LeftMenuLink from 'components/LeftMenuLink';
 
 import styles from './styles.scss';
 import messages from './messages.json';
 
-function LeftMenuLinkContainer({ plugins }) {
+function LeftMenuLinkContainer({ layout, plugins }) {
   const pluginsObject = plugins.toJS();
-
   // Generate the list of sections
   const pluginsSections = Object.keys(pluginsObject).reduce((acc, current) => {
     pluginsObject[current].leftMenuSections.forEach((section = {}) => {
       if (!isEmpty(section.links)) {
         acc[snakeCase(section.name)] = {
           name: section.name,
-          links: get(acc[snakeCase(section.name)], 'links', []).concat(section.links.map(link => {
-            link.source = current;
-            link.plugin = !isEmpty(pluginsObject[link.plugin]) ? link.plugin : pluginsObject[current].id;
+          links: get(acc[snakeCase(section.name)], 'links', []).concat(
+            section.links.map(link => {
+              link.source = current;
+              link.plugin = !isEmpty(pluginsObject[link.plugin])
+                ? link.plugin
+                : pluginsObject[current].id;
 
-            return link;
-          })),
+              return link;
+            }),
+          ),
         };
       }
     });
@@ -36,19 +39,34 @@ function LeftMenuLinkContainer({ plugins }) {
     return acc;
   }, {});
 
-  const linkSections = Object.keys(pluginsSections).map((current, j) => (
-    <div key={j}>
-      <p className={styles.title}>{pluginsSections[current].name}</p>
-      <ul className={styles.list}>
-        {sortBy(pluginsSections[current].links, 'label').map((link, i) =>
-          <LeftMenuLink key={`${i}-${link.label}`} icon={link.icon || 'caret-right'} label={link.label} destination={`/plugins/${link.plugin}/${link.destination}`} source={link.source} />
-        )}
-      </ul>
-    </div>
-  ));
+  const linkSections = Object.keys(pluginsSections).map((current, j) => {
+    const contentTypesToShow = get(layout, 'layout.contentTypesToShow');
+    const contentTypes = contentTypesToShow
+      ? pluginsSections[current].links.filter(
+        obj => findIndex(contentTypesToShow, ['destination', obj.destination]) !== -1,
+      )
+      : pluginsSections[current].links;
+
+    return (
+      <div key={j}>
+        <p className={styles.title}>{pluginsSections[current].name}</p>
+        <ul className={styles.list}>
+          {sortBy(contentTypes, 'label').map((link, i) => (
+            <LeftMenuLink
+              key={`${i}-${link.label}`}
+              icon={link.icon || 'caret-right'}
+              label={link.label}
+              destination={`/plugins/${link.plugin}/${link.destination}`}
+              source={link.source}
+            />
+          ))}
+        </ul>
+      </div>
+    );
+  });
 
   // Check if the plugins list is empty or not and display plugins by name
-  const pluginsLinks = !isEmpty(pluginsObject) ?
+  const pluginsLinks = !isEmpty(pluginsObject) ? (
     map(sortBy(pluginsObject, 'name'), plugin => {
       if (plugin.id !== 'email' && plugin.id !== 'content-manager') {
         return (
@@ -59,30 +77,29 @@ function LeftMenuLinkContainer({ plugins }) {
             destination={`/plugins/${get(plugin, 'id')}`}
           />
         );
-      }})
-    : (
-      <li className={styles.noPluginsInstalled}>
-        <FormattedMessage {...messages.noPluginsInstalled} />.
-      </li>
-    );
+      }
+    })
+  ) : (
+    <li className={styles.noPluginsInstalled}>
+      <FormattedMessage {...messages.noPluginsInstalled} />.
+    </li>
+  );
 
   return (
     <div className={styles.leftMenuLinkContainer}>
       {linkSections}
       <div>
-        <p className={styles.title}><FormattedMessage {...messages.plugins} /></p>
-        <ul className={styles.list}>
-          {pluginsLinks}
-        </ul>
+        <p className={styles.title}>
+          <FormattedMessage {...messages.plugins} />
+        </p>
+        <ul className={styles.list}>{pluginsLinks}</ul>
       </div>
       <div>
-        <p className={styles.title}><FormattedMessage {...messages.general} /></p>
+        <p className={styles.title}>
+          <FormattedMessage {...messages.general} />
+        </p>
         <ul className={styles.list}>
-          <LeftMenuLink
-            icon="list"
-            label={messages.listPlugins.id}
-            destination="/list-plugins"
-          />
+          <LeftMenuLink icon="list" label={messages.listPlugins.id} destination="/list-plugins" />
           <LeftMenuLink
             icon="shopping-basket"
             label={messages.installNewPlugin.id}
@@ -99,7 +116,12 @@ function LeftMenuLinkContainer({ plugins }) {
   );
 }
 
+LeftMenuLinkContainer.defaultProps = {
+  layout: {},
+};
+
 LeftMenuLinkContainer.propTypes = {
+  layout: PropTypes.object,
   plugins: PropTypes.object.isRequired,
 };
 
