@@ -19,10 +19,7 @@ import { get, includes, isFunction, map, omit } from 'lodash';
 import { compose } from 'redux';
 
 // Actions required for disabling and enabling the OverlayBlocker
-import {
-  disableGlobalOverlayBlocker,
-  enableGlobalOverlayBlocker,
-} from 'actions/overlayBlocker';
+import { disableGlobalOverlayBlocker, enableGlobalOverlayBlocker } from 'actions/overlayBlocker';
 
 import { pluginLoaded, updatePlugin } from 'containers/App/actions';
 import {
@@ -37,43 +34,47 @@ import { hideNotification } from 'containers/NotificationProvider/actions';
 // Design
 import ComingSoonPage from 'containers/ComingSoonPage';
 import Content from 'containers/Content';
+import LocaleToggle from 'containers/LocaleToggle';
+import CTAWrapper from 'components/CtaWrapper';
 import Header from 'components/Header/index';
-import HomePage from 'containers/HomePage';
-import InstallPluginPage from 'containers/InstallPluginPage';
+import HomePage from 'containers/HomePage/Loadable';
+import InstallPluginPage from 'containers/InstallPluginPage/Loadable';
 import LeftMenu from 'containers/LeftMenu';
-import ListPluginsPage from 'containers/ListPluginsPage';
+import ListPluginsPage from 'containers/ListPluginsPage/Loadable';
 import Logout from 'components/Logout';
-import NotFoundPage from 'containers/NotFoundPage';
+import NotFoundPage from 'containers/NotFoundPage/Loadable';
 import OverlayBlocker from 'components/OverlayBlocker';
-import PluginPage from 'containers/PluginPage';
+import PluginPage from 'containers/PluginPage/Loadable';
 
 // Utils
 import auth from 'utils/auth';
 import injectReducer from 'utils/injectReducer';
 import injectSaga from 'utils/injectSaga';
 
-import { getGaStatus } from './actions';
+import { getGaStatus, getLayout } from './actions';
 import reducer from './reducer';
 import saga from './saga';
 import selectAdminPage from './selectors';
 
 import styles from './styles.scss';
 
-export class AdminPage extends React.Component { // eslint-disable-line react/prefer-stateless-function
+const PLUGINS_TO_BLOCK_PRODUCTION = ['content-type-builder', 'settings-manager'];
+
+export class AdminPage extends React.Component {
+  // eslint-disable-line react/prefer-stateless-function
   state = { hasAlreadyRegistereOtherPlugins: false };
 
-  getChildContext = () => (
-    {
-      disableGlobalOverlayBlocker: this.props.disableGlobalOverlayBlocker,
-      enableGlobalOverlayBlocker: this.props.enableGlobalOverlayBlocker,
-      plugins: this.props.plugins,
-      updatePlugin: this.props.updatePlugin,
-    }
-  );
+  getChildContext = () => ({
+    disableGlobalOverlayBlocker: this.props.disableGlobalOverlayBlocker,
+    enableGlobalOverlayBlocker: this.props.enableGlobalOverlayBlocker,
+    plugins: this.props.plugins,
+    updatePlugin: this.props.updatePlugin,
+  });
 
   componentDidMount() {
     this.checkLogin(this.props);
     this.props.getGaStatus();
+    this.props.getLayout();
     ReactGA.initialize('UA-54313258-9');
   }
 
@@ -86,30 +87,52 @@ export class AdminPage extends React.Component { // eslint-disable-line react/pr
       }
     }
 
-    if (get(nextProps.plugins.toJS(), ['users-permissions', 'hasAdminUser']) !== get(this.props.plugins.toJS(), ['users-permissions', 'hasAdminUser'])) {
+    if (
+      get(nextProps.plugins.toJS(), ['users-permissions', 'hasAdminUser']) !==
+      get(this.props.plugins.toJS(), ['users-permissions', 'hasAdminUser'])
+    ) {
+      this.checkLogin(nextProps, true);
+    }
+
+    if (!this.hasUserPluginLoaded(this.props) && this.hasUserPluginLoaded(nextProps)) {
       this.checkLogin(nextProps);
     }
   }
 
-  checkLogin = (props) => {
+  checkLogin = (props, skipAction = false) => {
     if (props.hasUserPlugin && this.isUrlProtected(props) && !auth.getToken()) {
-      const endPoint = this.hasAdminUser(props) ? 'login': 'register';
+      if (!this.hasUserPluginLoaded(props)) {
+        return;
+      }
+
+      const endPoint = this.hasAdminUser(props) ? 'login' : 'register';
       this.props.history.push(`/plugins/users-permissions/auth/${endPoint}`);
     }
 
-    if (!this.isUrlProtected(props) && includes(props.location.pathname, 'auth/register') && this.hasAdminUser(props)) {
+    if (
+      !this.isUrlProtected(props) &&
+      includes(props.location.pathname, 'auth/register') &&
+      this.hasAdminUser(props) &&
+      !skipAction
+    ) {
       this.props.history.push('/plugins/users-permissions/auth/login');
     }
 
-    if (props.hasUserPlugin && !this.isUrlProtected(props) && !includes(props.location.pathname, 'auth/register') && !this.hasAdminUser(props)) {
+    if (
+      props.hasUserPlugin &&
+      !this.isUrlProtected(props) &&
+      !includes(props.location.pathname, 'auth/register') &&
+      !this.hasAdminUser(props)
+    ) {
       this.props.history.push('/plugins/users-permissions/auth/register');
     }
 
-    if (!props.hasUserPlugin || auth.getToken() && !this.state.hasAlreadyRegistereOtherPlugins) {
+    if (!props.hasUserPlugin || (auth.getToken() && !this.state.hasAlreadyRegistereOtherPlugins)) {
       map(omit(this.props.plugins.toJS(), ['users-permissions', 'email']), plugin => {
         switch (true) {
           case isFunction(plugin.bootstrap) && isFunction(plugin.pluginRequirements):
-            plugin.pluginRequirements(plugin)
+            plugin
+              .pluginRequirements(plugin)
               .then(plugin => {
                 return plugin.bootstrap(plugin);
               })
@@ -127,25 +150,54 @@ export class AdminPage extends React.Component { // eslint-disable-line react/pr
 
       this.setState({ hasAlreadyRegistereOtherPlugins: true });
     }
-  }
+  };
 
-  hasAdminUser = (props) => get(props.plugins.toJS(), ['users-permissions', 'hasAdminUser']);
+  hasUserPluginLoaded = props =>
+    typeof get(props.plugins.toJS(), ['users-permissions', 'hasAdminUser']) !== 'undefined';
 
-  isUrlProtected = (props) => !includes(props.location.pathname, get(props.plugins.toJS(), ['users-permissions', 'nonProtectedUrl']));
+  hasAdminUser = props => get(props.plugins.toJS(), ['users-permissions', 'hasAdminUser']);
+
+  isUrlProtected = props =>
+    !includes(props.location.pathname, get(props.plugins.toJS(), ['users-permissions', 'nonProtectedUrl']));
+
+  shouldDisplayLogout = () => auth.getToken() && this.props.hasUserPlugin && this.isUrlProtected(this.props);
 
   showLeftMenu = () => !includes(this.props.location.pathname, 'users-permissions/auth/');
 
+  retrievePlugins = () => {
+    const {
+      adminPage: { currentEnvironment },
+      plugins,
+    } = this.props;
+
+    if (currentEnvironment === 'production') {
+      let pluginsToDisplay = plugins;
+      PLUGINS_TO_BLOCK_PRODUCTION.map(plugin => (pluginsToDisplay = pluginsToDisplay.delete(plugin)));
+
+      return pluginsToDisplay;
+    }
+
+    return plugins;
+  };
+
   render() {
-    const leftMenu = this.showLeftMenu() ? <LeftMenu plugins={this.props.plugins} /> : '';
+    const { adminPage } = this.props;
     const header = this.showLeftMenu() ? <Header /> : '';
     const style = this.showLeftMenu() ? {} : { width: '100%' };
 
     return (
       <div className={styles.adminPage}>
-        {leftMenu}
-        { auth.getToken() && this.props.hasUserPlugin && this.isUrlProtected(this.props) ? (
-          <Logout />
-        ) : ''}
+        {this.showLeftMenu() && (
+          <LeftMenu
+            plugins={this.retrievePlugins()}
+            layout={adminPage.layout}
+            version={adminPage.strapiVersion}
+          />
+        )}
+        <CTAWrapper>
+          {this.shouldDisplayLogout() && <Logout />}
+          <LocaleToggle isLogged={this.shouldDisplayLogout() === true} />
+        </CTAWrapper>
         <div className={styles.adminPageRightWrapper} style={style}>
           {header}
           <Content {...this.props} showLeftMenu={this.showLeftMenu()}>
@@ -189,6 +241,7 @@ AdminPage.propTypes = {
   disableGlobalOverlayBlocker: PropTypes.func.isRequired,
   enableGlobalOverlayBlocker: PropTypes.func.isRequired,
   getGaStatus: PropTypes.func.isRequired,
+  getLayout: PropTypes.func.isRequired,
   hasUserPlugin: PropTypes.bool,
   history: PropTypes.object.isRequired,
   location: PropTypes.object.isRequired,
@@ -208,12 +261,27 @@ const mapStateToProps = createStructuredSelector({
 
 function mapDispatchToProps(dispatch) {
   return {
-    disableGlobalOverlayBlocker: () => { dispatch(disableGlobalOverlayBlocker()); },
-    enableGlobalOverlayBlocker: () => { dispatch(enableGlobalOverlayBlocker()); },
-    getGaStatus: () => { dispatch(getGaStatus()); },
-    onHideNotification: (id) => { dispatch(hideNotification(id)); },
-    pluginLoaded: (plugin) => { dispatch(pluginLoaded(plugin)); },
-    updatePlugin: (pluginId, updatedKey, updatedValue) => { dispatch(updatePlugin(pluginId, updatedKey, updatedValue)); },
+    disableGlobalOverlayBlocker: () => {
+      dispatch(disableGlobalOverlayBlocker());
+    },
+    enableGlobalOverlayBlocker: () => {
+      dispatch(enableGlobalOverlayBlocker());
+    },
+    getGaStatus: () => {
+      dispatch(getGaStatus());
+    },
+    getLayout: () => {
+      dispatch(getLayout());
+    },
+    onHideNotification: id => {
+      dispatch(hideNotification(id));
+    },
+    pluginLoaded: plugin => {
+      dispatch(pluginLoaded(plugin));
+    },
+    updatePlugin: (pluginId, updatedKey, updatedValue) => {
+      dispatch(updatePlugin(pluginId, updatedKey, updatedValue));
+    },
     dispatch,
   };
 }
@@ -222,10 +290,6 @@ const withConnect = connect(mapStateToProps, mapDispatchToProps);
 const withReducer = injectReducer({ key: 'adminPage', reducer });
 const withSaga = injectSaga({ key: 'adminPage', saga });
 
-export default compose(
-  withReducer,
-  withSaga,
-  withConnect,
-)(AdminPage);
+export default compose(withReducer, withSaga, withConnect)(AdminPage);
 
 // export default connect(mapStateToProps, mapDispatchToProps)(AdminPage);
