@@ -14,10 +14,10 @@ import getInjectors from './reducerInjectors';
 export default ({ key, reducer }) => (WrappedComponent) => {
   class ReducerInjector extends React.Component {
     static WrappedComponent = WrappedComponent;
+    static displayName = `withReducer(${(WrappedComponent.displayName || WrappedComponent.name || 'Component')})`;
     static contextTypes = {
       store: PropTypes.object.isRequired,
     };
-    static displayName = `withReducer(${(WrappedComponent.displayName || WrappedComponent.name || 'Component')})`;
 
     componentWillMount() {
       const { injectReducer } = this.injectors;
