@@ -18,10 +18,10 @@ import getInjectors from './sagaInjectors';
 export default ({ key, saga, mode }) => (WrappedComponent) => {
   class InjectSaga extends React.Component {
     static WrappedComponent = WrappedComponent;
+    static displayName = `withSaga(${(WrappedComponent.displayName || WrappedComponent.name || 'Component')})`;
     static contextTypes = {
       store: PropTypes.object.isRequired,
     };
-    static displayName = `withSaga(${(WrappedComponent.displayName || WrappedComponent.name || 'Component')})`;
 
     componentWillMount() {
       const { injectSaga } = this.injectors;
