@@ -13,17 +13,35 @@ module.exports = {
     try {
       ctx.send({ currentEnvironment: strapi.app.env });
     } catch(err) {
-      ctx.badRequest(null, [{ messages: [{ id: 'An error occured' }] }]);
+      ctx.badRequest(null, [{ messages: [{ id: 'An error occurred' }] }]);
     }
   },
 
-  getGaConfig: async ctx =>{
+  getStrapiVersion: async ctx => {
+    try {
+      const strapiVersion = _.get(strapi.config, 'info.strapi', null);
+      return ctx.send({ strapiVersion });
+    } catch(err) {
+      return ctx.badRequest(null, [{ messages: [{ id: 'The version is not available' }] }]);
+    }
+  },
+
+  getGaConfig: async ctx => {
     try {
       const allowGa = _.get(strapi.config, 'info.customs.allowGa', true);
       ctx.send({ allowGa });
     } catch(err) {
-      console.log(err)
-      ctx.badRequest(null, [{ messages: [{ id: 'An error occured' }] }]);
+      ctx.badRequest(null, [{ messages: [{ id: 'An error occurred' }] }]);
+    }
+  },
+
+  getLayout: async ctx => {
+    try {
+      const layout = require('../config/layout.js');
+
+      return ctx.send({ layout });
+    } catch(err) {
+      return ctx.badRequest(null, [{ messages: [{ id: 'An error occurred' }] }]);
     }
   },
 
@@ -36,14 +54,14 @@ module.exports = {
 
       strapi.log.info(`Installing ${plugin}...`);
 
-      exec(`node ${strapiBin} install ${plugin} ${port === '4000' ? '--dev' : ''}`);
+      exec(`node "${strapiBin}" install ${plugin} ${port === '4000' ? '--dev' : ''}`);
 
       ctx.send({ ok: true });
 
       strapi.reload();
     } catch(err) {
       strapi.reload.isWatching = true;
-      ctx.badRequest(null, [{ messages: [{ id: 'An error occured' }] }]);
+      ctx.badRequest(null, [{ messages: [{ id: 'An error occurred' }] }]);
     }
   },
 
@@ -57,7 +75,7 @@ module.exports = {
 
       ctx.send({ plugins });
     } catch(err) {
-      ctx.badRequest(null, [{ messages: [{ id: 'An error occured' }] }]);
+      ctx.badRequest(null, [{ messages: [{ id: 'An error occurred' }] }]);
     }
   },
 
@@ -69,14 +87,14 @@ module.exports = {
       strapi.reload.isWatching = false;
 
       strapi.log.info(`Uninstalling ${plugin}...`);
-      exec(`node ${strapiBin} uninstall ${plugin}`);
+      exec(`node "${strapiBin}" uninstall ${plugin}`);
 
       ctx.send({ ok: true });
 
       strapi.reload();
     } catch(err) {
       strapi.reload.isWatching = true;
-      ctx.badRequest(null, [{ messages: [{ id: 'An error occured' }] }]);
+      ctx.badRequest(null, [{ messages: [{ id: 'An error occurred' }] }]);
     }
   }
 };
