@@ -1,7 +1,7 @@
 'use strict';
 
 /**
- * Categoty.js service
+ * Category.js service
  *
  * @description: A set of functions similar to controller's actions to avoid code duplication.
  */
@@ -12,82 +12,134 @@ const _ = require('lodash');
 module.exports = {
 
   /**
-   * Promise to fetch all categoties.
+   * Promise to fetch all categories.
    *
    * @return {Promise}
    */
 
   fetchAll: (params) => {
-    const convertedParams = strapi.utils.models.convertParams('categoty', params);
+    // Convert `params` object to filters compatible with Mongo.
+    const filters = strapi.utils.models.convertParams('category', params);
+    // Select field to populate.
+    const populate = Category.associations
+      .filter(ast => ast.autoPopulate !== false)
+      .map(ast => ast.alias)
+      .join(' ');
 
-    return Categoty
+    return Category
       .find()
-      .where(convertedParams.where)
-      .sort(convertedParams.sort)
-      .skip(convertedParams.start)
-      .limit(convertedParams.limit)
-      .populate(_.keys(_.groupBy(_.reject(strapi.models.categoty.associations, {autoPopulate: false}), 'alias')).join(' '));
+      .where(filters.where)
+      .sort(filters.sort)
+      .skip(filters.start)
+      .limit(filters.limit)
+      .populate(populate);
   },
 
   /**
-   * Promise to fetch a/an categoty.
+   * Promise to fetch a/an category.
    *
    * @return {Promise}
    */
 
   fetch: (params) => {
-    return Categoty
-      .findOne(_.pick(params, _.keys(Categoty.schema.paths)))
-      .populate(_.keys(_.groupBy(_.reject(strapi.models.categoty.associations, {autoPopulate: false}), 'alias')).join(' '));
+    // Select field to populate.
+    const populate = Category.associations
+      .filter(ast => ast.autoPopulate !== false)
+      .map(ast => ast.alias)
+      .join(' ');
+
+    return Category
+      .findOne(_.pick(params, _.keys(Category.schema.paths)))
+      .populate(populate);
   },
 
   /**
-   * Promise to add a/an categoty.
+   * Promise to count categories.
+   *
+   * @return {Promise}
+   */
+
+  count: (params) => {
+    // Convert `params` object to filters compatible with Mongo.
+    const filters = strapi.utils.models.convertParams('category', params);
+
+    return Category
+      .count()
+      .where(filters.where);
+  },
+
+  /**
+   * Promise to add a/an category.
    *
    * @return {Promise}
    */
 
   add: async (values) => {
-    const data = await Categoty.create(_.omit(values, _.keys(_.groupBy(strapi.models.categoty.associations, 'alias'))));
-    await strapi.hook.mongoose.manageRelations('categoty', _.merge(_.clone(data), { values }));
-    return data;
+    // Extract values related to relational data.
+    const relations = _.pick(values, Category.associations.map(ast => ast.alias));
+    const data = _.omit(values, Category.associations.map(ast => ast.alias));
+
+    // Create entry with no-relational data.
+    const entry = await Category.create(data);
+
+    // Create relational data and return the entry.
+    return Category.updateRelations({ id: entry.id, values: relations });
   },
 
   /**
-   * Promise to edit a/an categoty.
+   * Promise to edit a/an category.
    *
    * @return {Promise}
    */
 
   edit: async (params, values) => {
-    // Note: The current method will return the full response of Mongo.
-    // To get the updated object, you have to execute the `findOne()` method
-    // or use the `findOneOrUpdate()` method with `{ new:true }` option.
-    await strapi.hook.mongoose.manageRelations('categoty', _.merge(_.clone(params), { values }));
-    return Categoty.update(params, values, { multi: true });
+    // Extract values related to relational data.
+    const relations = _.pick(values, Category.associations.map(a => a.alias));
+    const data = _.omit(values, Category.associations.map(a => a.alias));
+
+    // Update entry with no-relational data.
+    const entry = await Category.update(params, data, { multi: true });
+
+    // Update relational data and return the entry.
+    return Category.updateRelations(Object.assign(params, { values: relations }));
   },
 
   /**
-   * Promise to remove a/an categoty.
+   * Promise to remove a/an category.
    *
    * @return {Promise}
    */
 
   remove: async params => {
+    // Select field to populate.
+    const populate = Category.associations
+      .filter(ast => ast.autoPopulate !== false)
+      .map(ast => ast.alias)
+      .join(' ');
+
     // Note: To get the full response of Mongo, use the `remove()` method
     // or add spent the parameter `{ passRawResult: true }` as second argument.
-    const data = await Categoty.findOneAndRemove(params, {})
-      .populate(_.keys(_.groupBy(_.reject(strapi.models.categoty.associations, {autoPopulate: false}), 'alias')).join(' '));
+    const data = await Category
+      .findOneAndRemove(params, {})
+      .populate(populate);
 
-    _.forEach(Categoty.associations, async association => {
-      const search = (_.endsWith(association.nature, 'One')) ? { [association.via]: data._id } : { [association.via]: { $in: [data._id] } };
-      const update = (_.endsWith(association.nature, 'One')) ? { [association.via]: null } : { $pull: { [association.via]: data._id } };
+    if (!data) {
+      return data;
+    }
 
-      await strapi.models[association.model || association.collection].update(
-        search,
-        update,
-        { multi: true });
-    });
+    await Promise.all(
+      Category.associations.map(async association => {
+        const search = _.endsWith(association.nature, 'One') || association.nature === 'oneToMany' ? { [association.via]: data._id } : { [association.via]: { $in: [data._id] } };
+        const update = _.endsWith(association.nature, 'One') || association.nature === 'oneToMany' ? { [association.via]: null } : { $pull: { [association.via]: data._id } };
+
+        // Retrieve model.
+        const model = association.plugin ?
+          strapi.plugins[association.plugin].models[association.model || association.collection] :
+          strapi.models[association.model || association.collection];
+
+        return model.update(search, update, { multi: true });
+      })
+    );
 
     return data;
   }
