@@ -32,7 +32,7 @@ function EditRelations(props) {
       {map(filterRelationsUpload(props.schema.relations), (relation, key) => {
         if (relation.nature.toLowerCase().includes('morph') && relation[key]) return '';
 
-        const Select = ['oneWay', 'oneToOne', 'manyToOne', 'oneToManyMorph', 'oneToOneMorph'].includes(relation.nature) && relation.dominant ? SelectOne : SelectMany;
+        const Select = ['oneWay', 'oneToOne', 'manyToOne', 'oneToManyMorph', 'oneToOneMorph'].includes(relation.nature) ? SelectOne : SelectMany;
 
         return (
           <Select
