@@ -47,7 +47,7 @@ class TableHeader extends React.Component {
     });
 
     // Add empty th for actions column.
-    headers.push(<th key={`th_action`}></th>);
+    headers.push(<th key="th_action"></th>);
 
     return (
       <thead className={styles.tableHeader}>
