@@ -5,13 +5,28 @@
  */
 
 import {
+  ADD_FILTER,
   CHANGE_PARAMS,
   DELETE_DATA,
   DELETE_DATA_SUCCESS,
   GET_DATA,
   GET_DATA_SUCCEEDED,
+  ON_CHANGE,
+  ON_CLICK_REMOVE,
+  ON_TOGGLE_FILTERS,
+  OPEN_FILTERS_WITH_SELECTION,
+  REMOVE_ALL_FILTERS,
+  REMOVE_FILTER,
   SET_PARAMS,
+  SUBMIT,
 } from './constants';
+
+export function addFilter(filter) {
+  return {
+    type: ADD_FILTER,
+    filter,
+  };
+}
 
 export function changeParams({ target }) {
   return {
@@ -52,9 +67,58 @@ export function getDataSucceeded(data) {
   };
 }
 
-export function setParams(params) {
+export function onChange(index, key, value) {
+  return {
+    type: ON_CHANGE,
+    index,
+    key,
+    value,
+  };
+}
+
+export function onClickRemove(index) {
+  return {
+    type: ON_CLICK_REMOVE,
+    index,
+  };
+}
+
+export function openFiltersWithSelections(index) {
+  return {
+    type: OPEN_FILTERS_WITH_SELECTION,
+    index,
+  };
+}
+
+export function onToggleFilters() {
+  return {
+    type: ON_TOGGLE_FILTERS,
+  };
+}
+
+export function removeAllFilters() {
+  return {
+    type: REMOVE_ALL_FILTERS,
+  };
+}
+
+export function removeFilter(index) {
+  return {
+    type: REMOVE_FILTER,
+    index,
+  };
+}
+
+export function setParams(params, filters) {
   return {
     type: SET_PARAMS,
     params,
+    filters,
+  };
+}
+
+export function submit() {
+  return {
+    type: SUBMIT,
   };
 }
