@@ -1,8 +1,8 @@
 /**
-*
-* TableListRow
-*
-*/
+ *
+ * TableListRow
+ *
+ */
 
 import React from 'react';
 import PropTypes from 'prop-types';
@@ -14,8 +14,10 @@ import PopUpWarning from 'components/PopUpWarning';
 import styles from 'components/TableList/styles.scss';
 import { router } from 'app';
 /* eslint-disable jsx-a11y/no-static-element-interactions */
+/* eslint-disable react/jsx-curly-brace-presence */
 
-class TableListRow extends React.Component { // eslint-disable-line react/prefer-stateless-function
+class TableListRow extends React.Component {
+  // eslint-disable-line react/prefer-stateless-function
   constructor(props) {
     super(props);
     this.state = {
@@ -24,42 +26,69 @@ class TableListRow extends React.Component { // eslint-disable-line react/prefer
   }
 
   handleEdit = () => {
-    router.push(`/plugins/content-type-builder/#edit${this.props.rowItem.name}::contentType::baseSettings`);
-  }
+    router.push(
+      `/plugins/content-type-builder/#edit${this.props.rowItem.name}::contentType::baseSettings`,
+    );
+  };
 
-  handleDelete = (e) => {
+  handleDelete = e => {
     e.preventDefault();
     e.stopPropagation();
     this.props.onDelete(this.props.rowItem.name);
     this.setState({ showWarning: false });
-  }
+  };
 
   handleGoTo = () => {
-    router.push(`/plugins/content-type-builder/models/${this.props.rowItem.name}${this.props.rowItem.source ? `&source=${this.props.rowItem.source}`: ''}`);
-  }
+    router.push(
+      `/plugins/content-type-builder/models/${this.props.rowItem.name}${
+        this.props.rowItem.source ? `&source=${this.props.rowItem.source}` : ''
+      }`,
+    );
+  };
 
   toggleModalWarning = () => this.setState({ showWarning: !this.state.showWarning });
 
   handleShowModalWarning = () => this.setState({ showWarning: !this.state.showWarning });
 
   render() {
-    const pluginSource = this.props.rowItem.source ? <FormattedMessage id="content-type-builder.from">{(message) =><span style={{ fontStyle: 'italic', color: '#787E8F', fontWeight: '500' }}>({message}: {this.props.rowItem.source})</span>}</FormattedMessage> : '';
-    const temporary = this.props.rowItem.isTemporary ? <FormattedMessage id="content-type-builder.contentType.temporaryDisplay" /> : '';
-    const description = isEmpty(this.props.rowItem.description) ? '-' :  this.props.rowItem.description;
+    const pluginSource = this.props.rowItem.source ? (
+      <FormattedMessage id="content-type-builder.from">
+        {message => (
+          <span style={{ fontStyle: 'italic', color: '#787E8F', fontWeight: '500' }}>
+            ({message}: {this.props.rowItem.source})
+          </span>
+        )}
+      </FormattedMessage>
+    ) : (
+      ''
+    );
+    const temporary = this.props.rowItem.isTemporary ? (
+      <FormattedMessage id="content-type-builder.contentType.temporaryDisplay" />
+    ) : (
+      ''
+    );
+    const description = isEmpty(this.props.rowItem.description)
+      ? '-'
+      : this.props.rowItem.description;
     const spanStyle = this.props.rowItem.isTemporary ? '60%' : '100%';
-    const icons = this.props.rowItem.source ? [] : [{ icoType: 'pencil', onClick: this.handleEdit }, { icoType: 'trash', onClick: this.handleShowModalWarning }];
+    const icons = this.props.rowItem.source
+      ? []
+      : [
+        { icoType: 'pencil', onClick: this.handleEdit },
+        { icoType: 'trash', onClick: this.handleShowModalWarning },
+      ];
 
     return (
       <ListRow onClick={this.handleGoTo}>
         <div className={`col-md-4 ${styles.italic} ${styles.nameContainer}`}>
           <i className={`fa ${this.props.rowItem.icon}`} />
-          <span style={{ width: spanStyle }}>{startCase(this.props.rowItem.name)} &nbsp;{pluginSource}</span>
+          <span style={{ width: spanStyle }}>
+            {startCase(this.props.rowItem.name)} &nbsp;{pluginSource}
+          </span>
           &nbsp;{temporary}
         </div>
         <div className={`col-md-5 text-center ${styles.descriptionContainer}`}>
-          <div>
-            {description}
-          </div>
+          <div>{description}</div>
         </div>
         <div className="col-md-2 text-center">{this.props.rowItem.fields}</div>
         <div className="col-md-1">
