@@ -23,7 +23,6 @@ import ContentHeader from 'components/ContentHeader';
 import EmptyAttributesView from 'components/EmptyAttributesView';
 import Form from 'containers/Form';
 import List from 'components/List';
-import NoTableWarning from 'components/NoTableWarning';
 import PluginLeftMenu from 'components/PluginLeftMenu';
 
 import forms from 'containers/Form/forms.json';
@@ -35,7 +34,6 @@ import { storeData } from '../../utils/storeData';
 
 import {
   cancelChanges,
-  checkIfTableExists,
   deleteAttribute,
   modelFetch,
   modelFetchSucceeded,
@@ -50,9 +48,11 @@ import styles from './styles.scss';
 
 // Array of attributes that the ctb can handle at the moment
 const availableAttributes = Object.keys(forms.attribute);
+availableAttributes.push('integer', 'decimal', 'float');
 
 /* eslint-disable jsx-a11y/no-static-element-interactions */
 /* eslint-disable react/jsx-wrap-multilines */
+/* eslint-disable react/jsx-curly-brace-presence */
 
 export class ModelPage extends React.Component { // eslint-disable-line react/prefer-stateless-function
   constructor(props) {
@@ -83,30 +83,7 @@ export class ModelPage extends React.Component { // eslint-disable-line react/pr
         this.props.modelFetchSucceeded({ model: storeData.getContentType() });
       }
     }
-
-    if (this.props.modelPage.didFetchModel !== nextProps.modelPage.didFetchModel) {
-      this.props.checkIfTableExists();
-    }
   }
-
-  // componentWillUpdate(nextProps) {
-  //   if (!isEmpty(nextProps.menu)) {
-  //     const allowedPaths = nextProps.menu.reduce((acc, current) => {
-  //       const models = current.items.reduce((acc, current) => {
-  //         acc.push(current.name);
-  //
-  //         return acc;
-  //       }, []);
-  //       return acc.concat(models);
-  //     }, []);
-  //
-  //     const shouldRedirect = allowedPaths.filter(el => el === this.props.match.params.modelName.split('&')[0]).length === 0;
-  //
-  //     if (shouldRedirect) {
-  //       this.props.history.push('/404');
-  //     }
-  //   }
-  // }
 
   componentDidUpdate(prevProps) {
     if (prevProps.match.params.modelName !== this.props.match.params.modelName) {
@@ -283,7 +260,6 @@ export class ModelPage extends React.Component { // eslint-disable-line react/pr
     // Url to redirects the user if he modifies the temporary content type name
     const redirectRoute = replace(this.props.match.path, '/:modelName', '');
     const addButtons  = get(storeData.getContentType(), 'name') === this.props.match.params.modelName && size(get(storeData.getContentType(), 'attributes')) > 0 || this.props.modelPage.showButtons;
-    const showNoTableWarning = this.props.modelPage.tableExists ? '' : <NoTableWarning modelName={this.props.modelPage.model.name} />;
     const contentHeaderDescription = this.props.modelPage.model.description || 'content-type-builder.modelPage.contentHeader.emptyDescription.description';
     const content = size(this.props.modelPage.model.attributes) === 0 ?
       <EmptyAttributesView onClickAddAttribute={this.handleClickAddAttribute} /> :
@@ -321,7 +297,6 @@ export class ModelPage extends React.Component { // eslint-disable-line react/pr
 
                 />
                 {content}
-                {showNoTableWarning}
               </div>
             </div>
           </div>
@@ -350,9 +325,7 @@ ModelPage.contextTypes = {
 
 ModelPage.propTypes = {
   cancelChanges: PropTypes.func.isRequired,
-  checkIfTableExists: PropTypes.func.isRequired,
   deleteAttribute: PropTypes.func.isRequired,
-  // history: PropTypes.object.isRequired,
   location: PropTypes.object.isRequired,
   match: PropTypes.object.isRequired,
   menu: PropTypes.array.isRequired,
@@ -374,7 +347,6 @@ function mapDispatchToProps(dispatch) {
   return bindActionCreators(
     {
       cancelChanges,
-      checkIfTableExists,
       deleteAttribute,
       modelFetch,
       modelFetchSucceeded,
