@@ -1,6 +1,6 @@
 'use strict';
 
-const path = require('path')
+const path = require('path');
 const fs = require('fs');
 const _ = require('lodash');
 const generator = require('strapi-generate');
@@ -48,6 +48,10 @@ module.exports = {
 
     const pluginModels = Object.keys(strapi.plugins).reduce((acc, current) => {
       _.forEach(strapi.plugins[current].models, (model, name) => {
+        if (name === 'file') {
+          return true;
+        }
+
         acc.push({
           icon: 'fa-cube',
           name: _.get(model, 'info.name', 'model.name.missing'),
@@ -107,7 +111,7 @@ module.exports = {
     };
   },
 
-  getConnections: () => {
+  getConnections: () => {
     return _.keys(strapi.config.currentEnvironment.database.connections);
   },
 
@@ -121,7 +125,7 @@ module.exports = {
         rootPath: strapi.config.appPath,
         args: {
           api: name,
-          description: _.replace(description, /\"/g, '\\"'),
+          description: _.replace(description, /\"/g, '\\"'), // eslint-disable-line no-useless-escape
           attributes,
           connection,
           collectionName: !_.isEmpty(collectionName) ? collectionName : undefined,
@@ -177,7 +181,7 @@ module.exports = {
 
     _.forEach(attributesConfigurable, attribute => {
       if (_.has(attribute, 'params.type')) {
-        attrs[attribute.name] = attribute.params;
+        attrs[attribute.name] = _.omit(attribute.params, 'multiple');
 
         if (attribute.params.type === 'media') {
           const via = _.findKey(strapi.plugins.upload.models.file.attributes, {collection: '*'});
@@ -186,7 +190,7 @@ module.exports = {
             [attribute.params.multiple ? 'collection' : 'model']: 'file',
             via,
             plugin: 'upload'
-          }
+          };
         }
       } else if (_.has(attribute, 'params.target')) {
         const relation = attribute.params;
@@ -197,6 +201,7 @@ module.exports = {
         };
 
         switch (relation.nature) {
+          case 'oneWay':
           case 'oneToOne':
           case 'manyToOne':
             attr.model = relation.target;
@@ -208,9 +213,14 @@ module.exports = {
           default:
         }
 
-        attr.via = relation.key;
+        if(relation.nature !== 'oneWay') {
+          attr.via = relation.key;
+        }
         attr.dominant = relation.dominant;
-        attr.plugin = relation.pluginValue;
+
+        if (_.trim(relation.pluginValue)) {
+          attr.plugin = _.trim(relation.pluginValue);
+        }
 
         attrs[attribute.name] = attr;
       }
@@ -346,20 +356,32 @@ module.exports = {
             const attr = {};
 
             switch (params.nature) {
+              case 'oneWay':
+                return;
               case 'oneToOne':
               case 'oneToMany':
                 attr.model = model.toLowerCase();
                 break;
               case 'manyToOne':
-              case 'manyToMany':
                 attr.collection = model.toLowerCase();
                 break;
+              case 'manyToMany': {
+                attr.collection = model.toLowerCase();
+
+                if (!params.dominant) {
+                  attr.dominant = true;
+                }
+                break;
+              }
               default:
             }
 
             attr.via = name;
             attr.columnName = params.targetColumnName;
-            attr.plugin = source;
+
+            if (_.trim(source)) {
+              attr.plugin = _.trim(source);
+            }
 
             modelJSON.attributes[params.key] = attr;
 
@@ -442,7 +464,7 @@ module.exports = {
           }
         }
       }
-    }
+    };
 
     const recurciveDeleteFiles = folderPath => {
       try {
@@ -478,7 +500,7 @@ module.exports = {
           }
         });
       }
-    }
+    };
 
     recurciveDeleteFiles(apiPath);
 
