@@ -11,7 +11,7 @@ module.exports = {
       {
         name: 'menu.section.global-settings',
         items: [
-            {
+          {
             slug: 'application',
             name: 'menu.item.application',
             icon: 'globe'
@@ -105,7 +105,7 @@ module.exports = {
           ]
         }
       ]
-    }
+    };
   },
 
   request: env => ({
@@ -521,11 +521,6 @@ module.exports = {
                 port: 3306
               },
               {
-                name: 'form.database.item.provider.sqlite3',
-                value: 'sqlite3',
-                port: 1433
-              },
-              {
                 name: 'form.database.item.provider.redis',
                 value: 'redis',
                 port: 6379
@@ -575,7 +570,21 @@ module.exports = {
             type: 'password',
             value: _.get(strapi.config, `environments.${env}.database.connections.${name}.settings.password`, null),
             validations: {}
-          }
+          },
+          {
+            name: 'form.database.item.authenticationDatabase',
+            target: `database.connections.${name}.options.authenticationDatabase`,
+            type: 'string',
+            value: _.get(strapi.config, `environments.${env}.database.connections.${name}.options.authenticationDatabase`, null),
+            validations: {}
+          },
+          {
+            name: 'form.database.item.ssl',
+            target: `database.connections.${name}.options.ssl`,
+            type: 'boolean',
+            value: [true, 'true'].includes(_.get(strapi.config, `environments.${env}.database.connections.${name}.options.ssl`, false)),
+            validations: {}
+          },
         ]
       },
       {
@@ -583,7 +592,7 @@ module.exports = {
         items: [
           {
             name: 'form.database.item.default',
-            target: `database.defaultConnection`,
+            target: 'database.defaultConnection',
             type: 'string',
             value: _.get(strapi.config, `environments.${env}.database.defaultConnection`, null),
             validations: {
@@ -636,7 +645,7 @@ module.exports = {
   },
 
   getClientConnector: client => {
-    const bookshelfClients = ['postgres', 'mysql', 'sqlite3'];
+    const bookshelfClients = ['postgres', 'mysql'];
     const mongooseClients = ['mongo'];
     const redisClients = ['redis'];
 
@@ -661,9 +670,6 @@ module.exports = {
         break;
       case 'mongo':
         return '#43b121';
-        break;
-      case 'sqlite3':
-        return '#006fff';
         break;
       default:
         return '#000000';
@@ -885,8 +891,7 @@ module.exports = {
   installDependency: (params, name) => {
     const clientsDependencies = {
       postgres: 'pg',
-      mysql: 'mysql',
-      sqlite3: 'sqlite3'
+      mysql: 'mysql'
     };
 
     const client = _.get(clientsDependencies, _.get(params, `database.connections.${name}.settings.client`));
