@@ -59,11 +59,11 @@ export class HomePage extends React.Component {
 
   componentWillMount() {
     if (!isEmpty(this.props.location.search)) {
-      const page = parseInt(this.getURLParams('page'), 10);
-      const limit = parseInt(this.getURLParams('limit'), 10);
-      const sort = this.getURLParams('sort');
+      const _page = parseInt(this.getURLParams('_page'), 10);
+      const _limit = parseInt(this.getURLParams('_limit'), 10);
+      const _sort = this.getURLParams('_sort');
 
-      this.props.setParams({ limit, page, sort });
+      this.props.setParams({ _limit, _page, _sort });
     }
   }
   componentDidMount() {
@@ -82,12 +82,12 @@ export class HomePage extends React.Component {
   getURLParams = (type) => getQueryParameters(this.props.location.search, type);
 
   changeSort = (name) => {
-    const { params: { limit, page } } = this.props;
+    const { params: { _limit, _page } } = this.props;
     const target = {
-      name: 'params.sort',
+      name: 'params._sort',
       value: name,
     };
-    const search = `page=${page}&limit=${limit}&sort=${name}`;
+    const search = `_page=${_page}&_limit=${_limit}&_sort=${name}`;
 
     this.props.changeParams({ target });
     this.props.history.push({
@@ -98,9 +98,9 @@ export class HomePage extends React.Component {
 
   handleChangeParams = (e) => {
     const { history, params } = this.props;
-    const search = e.target.name === 'params.limit' ?
-      `page=${params.page}&limit=${e.target.value}&sort=${params.sort}`
-      : `page=${e.target.value}&limit=${params.limit}&sort=${params.sort}`;
+    const search = e.target.name === 'params._limit' ?
+      `_page=${params._page}&_limit=${e.target.value}&_sort=${params._sort}`
+      : `_page=${e.target.value}&_limit=${params._limit}&_sort=${params._sort}`;
     this.props.history.push({
       pathname: history.pathname,
       search,
@@ -142,12 +142,12 @@ export class HomePage extends React.Component {
         <div className={styles.entriesWrapper}>
           <div>
             {/* NOTE: Prepare for bulk actions}
-            <InputSelect
+              <InputSelect
               name="bulkAction"
               onChange={() => console.log('change')}
               selectOptions={[{ value: 'select all'}]}
               style={{ minWidth: '200px', height: '32px', marginTop: '-8px' }}
-            />
+              />
             */}
           </div>
           <EntriesNumber number={this.props.entriesNumber} />
@@ -155,7 +155,7 @@ export class HomePage extends React.Component {
         <List
           data={this.props.uploadedFiles}
           changeSort={this.changeSort}
-          sort={this.props.params.sort}
+          sort={this.props.params._sort}
         />
         <div className="col-md-12">
           <PageFooter
@@ -179,9 +179,9 @@ HomePage.contextTypes = {
 
 HomePage.defaultProps = {
   params: {
-    limit: 10,
-    page: 1,
-    sort: 'updatedAt',
+    _limit: 10,
+    _page: 1,
+    _sort: 'updatedAt',
   },
   uploadedFiles: [],
 };
