@@ -34,24 +34,24 @@ module.exports = {
     // Transform stream files to buffer
     const buffers = await strapi.plugins.upload.services.upload.bufferize(ctx.request.body.files.files);
     const enhancedFiles = buffers.map(file => {
-        if (file.size > config.sizeLimit) {
-          return ctx.badRequest(null, ctx.request.admin ? [{ messages: [{ id: 'Upload.status.sizeLimit', values: {file: file.name} }] }] : `${file.name} file is bigger than limit size!`);
-        }
+      if (file.size > config.sizeLimit) {
+        return ctx.badRequest(null, ctx.request.admin ? [{ messages: [{ id: 'Upload.status.sizeLimit', values: {file: file.name} }] }] : `${file.name} file is bigger than limit size!`);
+      }
 
-        // Add details to the file to be able to create the relationships.
-        if (refId && ref && field) {
-          Object.assign(file, {
-            related: [{
-              refId,
-              ref,
-              source,
-              field
-            }]
-          });
-        }
+      // Add details to the file to be able to create the relationships.
+      if (refId && ref && field) {
+        Object.assign(file, {
+          related: [{
+            refId,
+            ref,
+            source,
+            field
+          }]
+        });
+      }
 
-        return file;
-      });
+      return file;
+    });
 
     // Something is wrong (size limit)...
     if (ctx.status === 400) {
@@ -132,7 +132,7 @@ module.exports = {
     ctx.send(data);
   },
 
-  count: async (ctx, next) => {
+  count: async (ctx) => {
     const data = await strapi.plugins['upload'].services.upload.count(ctx.query);
 
     // Send 200 `ok`
@@ -141,7 +141,7 @@ module.exports = {
     });
   },
 
-  destroy: async (ctx, next) => {
+  destroy: async (ctx) => {
     const config = await strapi.store({
       environment: strapi.config.environment,
       type: 'plugin',
