@@ -4,12 +4,14 @@ module.exports = {
   find: async function (params = {}, populate) {
     const records = await this.query(function(qb) {
       _.forEach(params.where, (where, key) => {
-        qb.where(key, where[0].symbol, where[0].value);
+        if (_.isArray(where.value)) {
+          for (const value in where.value) {
+            qb[value ? 'where' : 'orWhere'](key, where.symbol, where.value[value]);
+          }
+        } else {
+          qb.where(key, where.symbol, where.value);
+        }
       });
-
-      if (params.sort) {
-        qb.orderByRaw(params.sort);
-      }
 
       if (params.start) {
         qb.offset(params.start);
@@ -18,26 +20,47 @@ module.exports = {
       if (params.limit) {
         qb.limit(params.limit);
       }
-    }).fetchAll({
-      withRelated: populate || _.keys(_.groupBy(_.reject(this.associations, { autoPopulate: false }), 'alias'))
-    });
+
+      if (params.sort) {
+        if (params.sort.key) {
+          qb.orderBy(params.sort.key, params.sort.order);
+        } else {
+          qb.orderBy(params.sort);
+        }
+      }
+    })
+      .fetchAll({
+        withRelated: populate || _.keys(_.groupBy(_.reject(this.associations, { autoPopulate: false }), 'alias'))
+      });
+
 
     return records ? records.toJSON() : records;
   },
 
   count: async function (params = {}) {
     return await this
-      .where(params)
+      .forge() // Instanciate the model
+      .query(qb => {
+        _.forEach(params.where, (where, key) => {
+          if (_.isArray(where.value)) {
+            for (const value in where.value) {
+              qb[value ? 'where' : 'orWhere'](key, where.symbol, where.value[value]);
+            }
+          } else {
+            qb.where(key, where.symbol, where.value);
+          }
+        });
+      })
       .count();
   },
 
   findOne: async function (params, populate) {
-    const primaryKey = params[this.primaryKey] || params.id;
+    const primaryKey = params[this.primaryKey] || params._id;
 
     if (primaryKey) {
       params = {
         [this.primaryKey]: primaryKey
-      }
+      };
     }
 
     const record = await this
@@ -85,7 +108,7 @@ module.exports = {
 
       search = {
         [this.primaryKey]: entry[this.primaryKey] || entry.id
-      }
+      };
     }
 
     return this.forge(search)
@@ -108,12 +131,20 @@ module.exports = {
       .destroy();
   },
 
+  deleteMany: async function (params) {
+    return await this
+      .query(qb => {
+        qb.whereIn(this.primaryKey, params[this.primaryKey] || params.id);
+      })
+      .destroy();
+  },
+
   search: async function (params) {
     return this
       .query(function(qb) {
         qb
-        .where('username', 'LIKE', `%${params.id}%`)
-        .orWhere('email', 'LIKE', `%${params.id}%`);
+          .where('username', 'LIKE', `%${params.id}%`)
+          .orWhere('email', 'LIKE', `%${params.id}%`);
       })
       .fetchAll();
   },
@@ -125,10 +156,13 @@ module.exports = {
   },
 
   removePermission: async function (params) {
+    const value = params[this.primaryKey] ? {
+      [this.primaryKey]: params[this.primaryKey] || params.id
+    } : params;
+
     return this
-      .forge({
-        [this.primaryKey]: params[this.primaryKey] || params.id
-      })
+      .forge()
+      .where(value)
       .destroy();
   }
 };
