@@ -6,8 +6,9 @@
  * @description: A set of functions called "actions" for managing `Auth`.
  */
 
-const _ = require('lodash');
+/* eslint-disable no-useless-escape */
 const crypto = require('crypto');
+const _ = require('lodash');
 const emailRegExp = /^(([^<>()\[\]\\.,;:\s@"]+(\.[^<>()\[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
 
 module.exports = {
@@ -142,7 +143,7 @@ module.exports = {
       }
     });
 
-    const provider = ctx.request.url.split('/')[2];
+    const provider = process.platform === 'win32' ? ctx.request.url.split('\\')[2] : ctx.request.url.split('/')[2];
     const config = grantConfig[provider];
 
     if (!_.get(config, 'enabled')) {
