@@ -79,7 +79,7 @@ module.exports = {
       const user = await strapi.query('user', 'users-permissions').findOne({ email: ctx.request.body.email });
 
       if (user) {
-        return ctx.badRequest(null, ctx.request.admin ? [{ messages: [{ id: 'Auth.form.error.email.taken' }] }] : 'Email is already taken.');
+        return ctx.badRequest(null, ctx.request.admin ? [{ messages: [{ id: 'Auth.form.error.email.taken', field: ['email'] }] }] : 'Email is already taken.');
       }
     }
 
@@ -98,7 +98,7 @@ module.exports = {
    * @return {Object}
    */
 
-  update: async (ctx, next) => {
+  update: async (ctx) => {
     try {
       const advancedConfigs = await strapi.store({
         environment: '',
@@ -111,13 +111,13 @@ module.exports = {
         const users = await strapi.plugins['users-permissions'].services.user.fetchAll({ email: ctx.request.body.email });
 
         if (users && _.find(users, user => (user.id || user._id).toString() !== ctx.params.id)) {
-          return ctx.badRequest(null, ctx.request.admin ? [{ messages: [{ id: 'Auth.form.error.email.taken' }] }] : 'Email is already taken.');
+          return ctx.badRequest(null, ctx.request.admin ? [{ messages: [{ id: 'Auth.form.error.email.taken', field: ['email'] }] }] : 'Email is already taken.');
         }
       }
 
       const user = await strapi.plugins['users-permissions'].services.user.fetch(ctx.params);
 
-      if (_.get(ctx.request, 'body.password') === user.password) {
+      if (_.get(ctx.request, 'body.password') === user.password) {
         delete ctx.request.body.password;
       }
 
@@ -131,7 +131,7 @@ module.exports = {
         });
 
         if (user !== null && (user.id || user._id).toString() !== ctx.params.id) {
-          return ctx.badRequest(null, ctx.request.admin ? [{ messages: [{ id: 'Auth.form.error.email.taken' }] }] : 'Email is already taken.');
+          return ctx.badRequest(null, ctx.request.admin ? [{ messages: [{ id: 'Auth.form.error.email.taken', field: ['email'] }] }] : 'Email is already taken.');
         }
       }
 
@@ -150,7 +150,7 @@ module.exports = {
    * @return {Object}
    */
 
-  destroy: async (ctx, next) => {
+  destroy: async (ctx) => {
     const data = await strapi.plugins['users-permissions'].services.user.remove(ctx.params);
 
     // Send 200 `ok`
